Split the chunks into two classes. One for annotated chunks (with speaker information added) and for regular chunks (just sound data)
